# api/openapi.yaml
openapi: 3.0.0
info:
  title: Booking Service API
  description: Simplified Cal.com clone — booking service API
  version: 1.0.0
tags: []
paths:
  /bookings:
    post:
      operationId: Bookings_createBooking
      description: Book a slot. Returns 409 if the slot is already taken
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreate'
    get:
      operationId: Bookings_listBookings
      description: List all upcoming bookings sorted by startTime
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
  /event-types:
    post:
      operationId: EventTypes_createEventType
      description: Create a new event type
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventTypeCreate'
    get:
      operationId: EventTypes_listEventTypes
      description: List all event types
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventType'
  /event-types/{id}:
    get:
      operationId: EventTypes_getEventType
      description: Get event type by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      operationId: EventTypes_patchEventType
      description: Partially update an event type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventTypePatch'
  /event-types/{id}/slots:
    get:
      operationId: EventTypes_getSlots
      description: Get available slots for the next 14 days
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Slot'
components:
  schemas:
    Booking:
      type: object
      required:
        - id
        - eventTypeId
        - startTime
        - endTime
        - guestName
        - guestEmail
        - createdAt
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        eventTypeId:
          $ref: '#/components/schemas/Uuid'
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        guestName:
          type: string
        guestEmail:
          type: string
        createdAt:
          type: string
          format: date-time
      description: A booking made by a guest
    BookingCreate:
      type: object
      required:
        - eventTypeId
        - startTime
        - guestName
        - guestEmail
      properties:
        eventTypeId:
          $ref: '#/components/schemas/Uuid'
        startTime:
          type: string
          format: date-time
        guestName:
          type: string
        guestEmail:
          type: string
      description: Payload to create a new booking
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
      description: Error response returned on 4xx failures
    EventType:
      type: object
      required:
        - id
        - title
        - description
        - durationMinutes
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        title:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
      description: Type of event that can be booked
    EventTypeCreate:
      type: object
      required:
        - title
        - description
        - durationMinutes
      properties:
        title:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
      description: Payload to create a new event type
    EventTypePatch:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
      description: Payload to patch an existing event type (all fields optional)
    ServiceVersions:
      type: string
      enum:
        - 1.0.0
    Slot:
      type: object
      required:
        - startTime
        - endTime
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
      description: Available time slot for booking
    Uuid:
      type: string
      format: uuid
      description: UUID identifier